#632 - move changes list to shared
