+++ src/components/shared/changes-list/changes-list.tsx
@@ -6,10 +6,10 @@
 import { Link } from 'react-router-dom';
 import { OverlayTrigger, Tooltip } from 'react-bootstrap';
 import { ImplementedChange } from 'utils';
-import { routes } from '../../../../../../shared/routes';
-import BulletList from '../../../../../shared/bullet-list/bullet-list';
+import { fullNamePipe, datePipe } from '../../../shared/pipes';
+import { routes } from '../../../shared/routes';
+import BulletList from '../bullet-list/bullet-list';
 import './changes-list.module.css';
-import { datePipe, fullNamePipe } from '../../../../../../shared/pipes';
 
 interface ChangesListProps {
   changes: ImplementedChange[];
